+++ modules/oracle/depinject.go
@@ -26,7 +26,8 @@ func (am AppModule) IsOnePerModuleType() {}
 // IsAppModule implements the appmodule.AppModule interface.
 func (am AppModule) IsAppModule() {}
 
-type OracleInputs struct {
+// Inputs define the module inputs for the depinject.
+type Inputs struct {
 	depinject.In
 
 	Config *modulev1.Module
@@ -38,14 +39,22 @@ type OracleInputs struct {
 	ServiceKeeper types.ServiceKeeper
 }
 
-type OracleOutputs struct {
+// Outputs define the module outputs for the depinject.
+type Outputs struct {
 	depinject.Out
 
 	OracleKeeper keeper.Keeper
 	Module       appmodule.AppModule
 }
 
-func ProvideModule(in OracleInputs) OracleOutputs {
+// ProvideModule creates a new OracleKeeper and AppModule using the provided inputs and returns the Outputs.
+//
+// Parameters:
+// - in: the Inputs struct containing the necessary dependencies for creating the OracleKeeper and AppModule.
+//
+// Returns:
+// - Outputs: the struct containing the OracleKeeper and AppModule.
+func ProvideModule(in Inputs) Outputs {
 	keeper := keeper.NewKeeper(
 		in.Cdc,
 		in.Key,
@@ -53,5 +62,5 @@ func ProvideModule(in OracleInputs) OracleOutputs {
 	)
 	m := NewAppModule(in.Cdc, keeper, in.AccountKeeper, in.BankKeeper)
 
-	return OracleOutputs{OracleKeeper: keeper, Module: m}
+	return Outputs{OracleKeeper: keeper, Module: m}
 }
